requirements.txt
# FastAPI Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Environment Management  
python-dotenv==1.0.0

# Google Cloud
google-cloud-bigquery==3.13.0
google-cloud-core==2.4.1
google-auth==2.25.2
google-cloud-logging

# HTTP Requests
httpx==0.25.2
requests==2.31.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black
flake8

# Production WSGI Server
gunicorn==21.2.0

# Data Validation
pydantic==2.5.0
python-multipart

# E-Mail Integration
beautifulsoup4
apscheduler
imapclient==2.3.1


